Started refinements to lambda-evaluator in order to do  a proper validation :
Added confidence intervals evaluation for the mean of the lambdas the evaluator generates, evaluated standard deviation, too.

Moved lambda-evaluator as a submodule of computational-model

Extended json-template, deserializer and model to include, for every ride, the expected wait time.

Added in docs a new file to report the confidence intervals evaluations

Co-authored-by: Alessandro Lioi <lioia@users.noreply.github.com>